# docs/openapi.yaml
openapi: 3.0.3
info:
  title: CashLenX API
  description: Personal finance management API
  version: 0.3.0
  contact:
    name: CashLenX Team
    email: support@cashlenx.com
servers:
  - url: http://localhost:8080
    description: Local development server
paths:
...
  /api/cash/expense:
    post:
      summary: Create expense transaction
      description: Record a new expense transaction
      operationId: createExpense
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CashFlowRequest'
      responses:
        '201':
          description: Expense transaction created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseWrapper'
  /api/cash/income:
    post:
      summary: Create income transaction
      description: Record a new income transaction
      operationId: createIncome
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CashFlowRequest'
      responses:
        '201':
          description: Income transaction created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseWrapper'
...
  /api/cash/summary/daily/{date}:
    get:
      summary: Get daily summary
      description: Get financial summary for a specific date
      operationId: getDailySummary
      parameters:
        - name: date
          in: path
          required: true
          schema:
            type: string
            pattern: '^\d{8}$'
      responses:
        '200':
          description: Daily summary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseWrapper'
  /api/cash/summary/monthly/{month}:
    get:
      summary: Get monthly summary
      description: Get financial summary for a specific month (YYYYMM format)
      operationId: getMonthlySummary
      parameters:
        - name: month
          in: path
          required: true
          schema:
            type: string
            pattern: '^\d{6}$'
      responses:
        '200':
          description: Monthly summary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseWrapper'
  /api/cash/summary/yearly/{year}:
    get:
      summary: Get yearly summary
      description: Get financial summary for a specific year (YYYY format)
      operationId: getYearlySummary
      parameters:
        - name: year
          in: path
          required: true
          schema:
            type: string
            pattern: '^\d{4}$'
      responses:
        '200':
          description: Yearly summary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseWrapper'

  # Category endpoints
...
  /api/manage/restore:
    post:
      summary: Restore database from dump
      description: Restore the database from a JSON dump file
      operationId: restoreDatabase
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: string
              format: binary
      responses:
        '200':
          description: Database restored successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseWrapper'
  /api/manage/export:
    get:
      summary: Export data to Excel
      description: Export cash flow data to an Excel file
      operationId: exportData
      parameters:
        - name: from
          in: query
          description: Start date for export (YYYYMMDD format, optional for all data)
          schema:
            type: string
            pattern: '^\d{8}$'
        - name: to
          in: query
          description: End date for export (YYYYMMDD format, optional for all data)
          schema:
            type: string
            pattern: '^\d{8}$'
      responses:
        '200':
          description: Excel file with exported data
          content:
            application/vnd.openxmlformats-officedocument.spreadsheetml.sheet:
              schema:
                type: string
                format: binary
  /api/manage/import:
    post:
      summary: Import data from Excel
      description: Import cash flow data from an Excel file
      operationId: importData
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                  description: Excel file containing cash flow data
              required:
                - file
      responses:
        '200':
          description: Data imported successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseWrapper'

  # User management endpoints
...
components:
  schemas:
...
    ResponseWrapper:
      type: object
      properties:
        code:
          type: string
          description: Response code for machine consumption
        message:
          type: string
          description: Human-readable message
        data:
          description: The actual data payload
          nullable: true
          oneOf:
            - type: object
            - type: array
              items: {}
            - type: string
            - type: number
            - type: boolean
        meta:
          $ref: '#/components/schemas/MetaInfo'
        extra:
          type: object
          description: Additional payload for non-standard fields
          additionalProperties: true
        errors:
          type: array
          description: Structured error list (empty for success)
          items:
            $ref: '#/components/schemas/ErrorItem'
      required:
        - code
        - data
        - message
        - meta
        - extra
        - errors
    ErrorItem:
      type: object
      properties:
        field:
          type: string
          description: The specific field that caused the error (optional)
        message:
          type: string
          description: Human-readable error message
      required:
        - message
    MetaInfo:
      type: object
      properties:
        total_count:
          type: integer
          description: Total number of items (for pagination)
        limit:
          type: integer
          description: Items per page
        offset:
          type: integer
          description: Offset for pagination
        request_id:
          type: string
          description: Unique request identifier

    # Cash Flow schemas
    CashFlowRequest:
      type: object
      properties:
        belongs_date:
          type: string
          description: "Transaction date (supported formats: YYYYMMDD, YYYY-MM-DD, YYYY/MM/DD)"
          pattern: ^(\d{8}|\d{4}[-/]\d{2}[-/]\d{2})$
        category_name:
          type: string
          description: Category name for the transaction
        amount:
          type: number
          format: double
          description: Transaction amount
        description:
          type: string
          description: Transaction description
      required:
        - belongs_date
        - category_name
        - amount
...